### 2016-10-06 #84 @arfon
> I also am not sure how well load-balanced we are - I currently am reviewing one, and have one more that I am the assigned editor for.

👍 ok no problem. @openjournals/joss-editors does anyone feel like they could handle this submission?
