### 2016-10-14 #93 @whedon
[REVIEW]: Golo - a lightweight dynamic language for the JVM
**Submitting author:** @jponge (<a href="http://orcid.org/0000-0003-3452-9183">Julien Ponge</a>)
**Repository:** <a href="https://github.com/eclipse/golo-lang" target ="_blank">https://github.com/eclipse/golo-lang</a>
**Version:** v3.2.0
