### 2016-10-14 #93 @danielskatz
Thanks @stevenrbrandt for agreeing to do this - the review is this issue.

The reviewer guidelines are here: http://joss.theoj.org/about#reviewer_guidelines

For the most part you need to see if the statements above in the first post are true and check things off accordingly. If something isn't the case, or you otherwise have a suggestion, just post a comment to @jponge here (or ask me) and we'll sort things out until the submission meets all of JOSS' guidelines.
