### 2016-10-17 #93 @jponge
I've made some updates, see @arfon's feedback https://github.com/eclipse/golo-lang/pull/410
