### 2016-10-19 #93 @stevenrbrandt
I'm having difficulty with the statement of need. All I see is this: "The world didn't need another JVM language. So we built yet another one. A simple one." I see a better statement of need here (http://www.javamagazine.mozaicreader.com/MarApr2016/Twitter#&pageSet=54&page=0), but I think it ought to go somewhere on the website or in the documentation.
