### 2016-10-24 #93 @danielskatz
@stevenrbrandt - Is this all of your feedback for the authors to respond to?  Or should they wait for more feedback?
