### 2016-10-24 #93 @stevenrbrandt
OK, here's a new issue. My command-line tests worked, but I am unable to get the Eclipse plugin to function. Not sure if that's my fault somehow. I filed an issue at https://github.com/golo-lang/gldt/issues/4
