### 2016-10-24 #93 @jponge
@stevenrbrandt the Eclipse plugin is (ironically) unmaintained at this stage and is not part of the submission.
