### 2016-10-27 #93 @stevenrbrandt
One of the purposes of golo is to teach people how to implement a dynamic runtime language (as stated in the paper). Given that, it seems to me there should be some document introducing the overall structure of the source code, maybe some tutorial on how to add a language feature. I don't see these things as blocking issues, however, just suggestions.
