### 2016-10-27 #93 @stevenrbrandt
@jponge, given that the Eclipse plugin isn't maintained and isn't working, it should be taken off the list of editors.
