### 2016-10-27 #93 @stevenrbrandt
shebang doesn't work for me. All the other methods for compiling and running this same source file do work
`$ ./build/install/golo/bin/golo shebang hello-1.golo
[error] In Golo module: duplicated-struct.golo
[error] Declaring a type `Foo` twice (declared first here: {line=7, column=1}) at {line=9, column=1}
`
