### 2016-11-02 #93 @jponge
@stevenrbrandt we are not removing the reference to the Eclipse IDE plugin, it keeps contributions hopes alive.

Regarding the `shebang` issue it seems to me that you are running `duplicated-struct.golo` from our tests suite and get an error. This is fine, this file is used by the test suite to check that an error is actually being detected at compilation time.
