### 2016-11-02 #93 @stevenrbrandt
@jponge How do I prevent the running of the duplicate struct-golo? I'm not explicitly doing that.
