### 2016-11-02 #93 @jponge
Could you please share the code snippet that raises the error, and if it comes from the Golo source code point me to the file path?

I do suspect that you are running a file that is actually meant to raise an error as part of our unit tests.
