### 2016-11-02 #93 @stevenrbrandt
I copied from the tutorial.
`module javamag.Hello

import java.util

# A comment
function main = |args| {
  let elements = ArrayList()
  elements: add("Hello")
  elements: add("World")
  elements: add("!")
  let size = elements:size()
  for(var i=0, i < size, i = i + 1) {
    print(elements:get(i))
    if(i < size - 1) {
      print(" ")
    }
  }
  println("")
}
`
