### 2016-11-02 #93 @stevenrbrandt
@jponge Nope. I don't get that result. Not sure what's different about my installation.
I do these things:
git clone https://github.com/eclipse/golo-lang
./gradlew build
./gradlew test
./gradlew clean
./gradlew installDist
./build/install/golo/bin/golo shebang hello-1.golo

And I get that error message I sent to you.

In case it's useful
$ uname -a
Linux localhost.localdomain 4.4.14-200.fc22.x86_64 #1 SMP Fri Jun 24 21:19:33 UTC 2016 x86_64 x86_64 x86_64 GNU/Linux
$ java -version
java version "1.8.0_65"
Java(TM) SE Runtime Environment (build 1.8.0_65-b17)
Java HotSpot(TM) 64-Bit Server VM (build 25.65-b01, mixed mode)
