### 2016-11-03 #93 @jponge
I'm on a fresher version of the JVM:

```
$ java -version
java version "1.8.0_112"
Java(TM) SE Runtime Environment (build 1.8.0_112-b16)
Java HotSpot(TM) 64-Bit Server VM (build 25.112-b16, mixed mode)
```

I did the same steps, here's the transcript.

```
$ git clone https://github.com/eclipse/golo-lang
Cloning into 'golo-lang'...
remote: Counting objects: 24291, done.
remote: Total 24291 (delta 4), reused 0 (delta 0), pack-reused 24279
Resolving deltas: 100% (9428/9428), done.
Checking connectivity... done.
fuji:Code jponge$ cd golo-lang/
```

Then I did a `./gradlew build installDist`, which builds just fine.

I then went with `check.golo` as follows:

```golo
module javamag.Hello

import java.util

# A comment
function main = |args| {
  let elements = ArrayList()
  elements: add("Hello")
  elements: add("World")
  elements: add("!")
  let size = elements:size()
  for(var i=0, i < size, i = i + 1) {
    print(elements:get(i))
    if(i < size - 1) {
      print(" ")
    }
  }
  println("")
}
```

It runs fine:

```
$ build/install/golo/bin/golo golo --files check.golo
Hello World !
```
