### 2016-11-03 #93 @stevenrbrandt
@jponge , that's not the shebang command. The files command works for me.

$ build/install/golo/bin/golo golo --files hello-1.golo
Hello World !
$ build/install/golo/bin/golo shebang hello-1.golo
[error] In Golo module: duplicated-struct.golo
[error] Declaring a type `Foo` twice (declared first here: {line=7, column=1}) at {line=9, column=1}
