### 2016-11-03 #93 @jponge
Ok I think I understand now :-)

 shebang is special as it scans for all .golo and .jar files starting
 from your current working directory.

You should try to move the file and you current working directory
somewhere else then. Does it work for you?
