### 2016-11-03 #93 @stevenrbrandt
@jponge, Yes, putting it in its own directory does make a difference. That point is mentioned in the docs, but I didn't recognize that my problem might be caused by that feature. Maybe you could elaborate on this feature in the docs to make it clearer why golo scans the subdirectories and what functionality that enables. Also, you might mention the problems it might create in the pitfalls section.
