### 2016-11-04 #93 @jponge
Good point, I'll think about adding this in the documentation.
