### 2016-11-10 #93 @danielskatz
@stevenrbrandt - it looks like the two unchecked boxes are
Functionality: Have the functional claims of the software been confirmed?
Performance: Have any performance claims of the software been confirmed?

Will the documentation fix that you suggested allow you to check the functionality box?

If the software doesn't make performance claims, this box can be checked.
If it does make performance claims, then those need to be reviewed before the box can be checked.
