### 2016-11-17 #93 @jponge
@danielskatz I've just pushed documentation changes to address @stevenrbrandt feedback, see https://github.com/eclipse/golo-lang/commit/6167b7122a466820716e4068a4419bca07b57375
