### 2016-11-28 #93 @danielskatz
Ok, thanks - I see that everything is now checked.

@arfon, this one is ready to accept! :)
