### 2016-11-29 #93 @arfon
> @arfon, this one is ready to accept! :)

Excellent. Thanks @danielskatz.
