### 2016-12-02 #93 @arfon
@stevenrbrandt many thanks for the review. @danielskatz thanks for editing this submission.

@jponge your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00093 🚀 🎉 💥
