### 2016-12-02 #93 @jponge
Thanks everyone :-)

On Fri, Dec 2, 2016 at 1:52 AM, Arfon Smith <notifications@github.com>
wrote:

> Closed #93 <https://github.com/openjournals/joss-reviews/issues/93>.
>
